show user detail on click

+++ app/src/main/java/com/giftech/dicodinguser/ui/navigation/Screen.kt
@@ -4,7 +4,7 @@ package com.giftech.dicodinguser.ui.navigation
 sealed class Screen(val route: String) {
     object Home : Screen("home")
     object DetailUser : Screen("detail/{userId}") {
-        fun createRoute(userId: Long) = "detail/$userId"
+        fun createRoute(userId: Int) = "detail/$userId"
     }
     object Aboout : Screen("about")
 }
